requirements.txt
matplotlib==3.8.0
numpy==1.26.0
pillow==10.0.1
pytorch==2.0.0
pyyaml==6.0.1
torchvision==0.15.0